# app/swagger.json
{
  "swagger": "2.0",
  "info": {
    "contact": {},
    "description": "This is an implemnetation for the Interview challenge.",
    "title": "Interview challenge",
    "version": "1.0"
  },
  "basePath": "/v1",
  "paths": {
    "/history": {
      "get": {
        "consumes": [
          "application/json"
        ],
        "description": "List queries",
        "operationId": "queries_history",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "items": {
                "properties": {
                  "addresses": {
                    "items": {
                      "properties": {
                        "ip": {
                          "type": "string"
                        }
                      },
                      "type": "object"
                    },
                    "type": "array"
                  },
                  "client_ip": {
                    "type": "string"
                  },
                  "created_at": {
                    "type": "integer"
                  },
                  "domain": {
                    "type": "string"
                  }
                },
                "type": "object"
              },
              "type": "array"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "properties": {
                "message": {
                  "type": "string"
                }
              },
              "type": "object"
            }
          }
        },
        "summary": "List queries",
        "tags": [
          "history"
        ]
      }
    },
    "/tools/lookup": {
      "get": {
        "consumes": [
          "application/json"
        ],
        "description": "Lookup domain and return all IPv4 addresses",
        "operationId": "lookup_domain",
        "parameters": [
          {
            "description": "Domain name",
            "in": "query",
            "name": "domain",
            "required": true,
            "type": "string"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "properties": {
                "addresses": {
                  "items": {
                    "properties": {
                      "ip": {
                        "type": "string"
                      }
                    },
                    "type": "object"
                  },
                  "type": "array"
                },
                "client_ip": {
                  "type": "string"
                },
                "created_at": {
                  "type": "integer"
                },
                "domain": {
                  "type": "string"
                }
              },
              "type": "object"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "properties": {
                "message": {
                  "type": "string"
                }
              },
              "type": "object"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "properties": {
                "message": {
                  "type": "string"
                }
              },
              "type": "object"
            }
          }
        },
        "summary": "Lookup domain",
        "tags": [
          "tools"
        ]
      }
    },
    "/tools/validate": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "description": "Simple IP valication",
        "operationId": "validate_ip",
        "parameters": [
          {
            "description": "IP to validate",
            "in": "body",
            "name": "request",
            "required": true,
            "schema": {
              "properties": {
                "ip": {
                  "type": "string"
                }
              },
              "type": "object"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "properties": {
                "status": {
                  "type": "boolean"
                }
              },
              "type": "object"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "properties": {
                "message": {
                  "type": "string"
                }
              },
              "type": "object"
            }
          }
        },
        "summary": "Simple IP validation",
        "tags": [
          "tools"
        ]
      }
    }
  },
  "definitions": {
    "ValidateIPRequest": {
      "properties": {
        "ip": {
          "type": "string"
        }
      },
      "type": "object"
    },
    "ValidateIPResponse": {
      "properties": {
        "status": {
          "type": "boolean"
        }
      },
      "type": "object"
    },
    "Address": {
      "properties": {
        "ip": {
          "type": "string"
        }
      },
      "type": "object"
    },
    "Query": {
      "properties": {
        "addresses": {
          "items": {
            "properties": {
              "ip": {
                "type": "string"
              }
            },
            "type": "object"
          },
          "type": "array"
        },
        "client_ip": {
          "type": "string"
        },
        "created_at": {
          "type": "integer"
        },
        "domain": {
          "type": "string"
        }
      },
      "type": "object"
    },
    "HTTPError": {
      "properties": {
        "message": {
          "type": "string"
        }
      },
      "type": "object"
    }
  }
}